\name{calculate.rho.signed}
\alias{calculate.rho.signed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
correlation calculation
}
\description{
correlation analysis with FDR calculation
}
\usage{
calculate.rho.signed(datExpr,n.perm = 10,FDR.cutoff = 0.05,estimator = "pearson",
                               use.obs = "na.or.complete",
                               direction = "absolute",
                               rho.thresh = NULL,sort.el = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datExpr}{gene expression data matrix}
  \item{n.perm}{Number of permutations to perform. If \code{doPerm = NULL}, calculates BH FDR p-values instead of permutation based FDR.}
  \item{FDR.cutoff}{FDR threshold to output final results of significant correlations.}
  \item{estimator}{Type of correlation coefficient to calculate, and gets passed to 'method' in 'cor()' function. Either "pearson" or "spearman".}
  \item{direction}{If direction = "absolute", absolute valued correlations are considered. If direction = "positive" or "negative", signed correlation is considered, and significance is calculated for positive/negative values.}
  \item{rho.thresh}{vector of correlation cutoff threshold to calculate significance. If NULL, threshold values will be generated by increment of 0.01 between -1 and 1 (if direction = "positive" or "negative"), or 0 and 1 (if direction = "absolute").}
  \item{sort.el}{logical value to determine whether correlation list should be sorted}
  \item{use.obs}{an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}
}
\details{
 direction = "absolute" can serve as two-tailed significance test, where as direction = "positive"/"negative" is one-tailed test. 
}
\value{
 output is a list. output$signif.ijw is three column edgelist data.frame, third column being the weight. output$FDR is the permutation FDR table across the threshold values in rho.thresh.
}
\author{
Won-Min Song
}
\examples{

data(Sample_Expression)
rho.out = calculate.rho.signed(datExpr,n.perm = 10,FDR.cutoff = 0.05,estimator = "pearson",
                               use.obs = "na.or.complete",
                               direction = "absolute",
                               rho.thresh = NULL,sort.el = TRUE)

}