\name{datExpr}
\docType{data}
\alias{datExpr}
\title{
Toy example data
}
\description{
A portion of TCGA breast cancer data to test run MEGENA. 
}
\usage{
Sample_Expression.RData
}
\format{
Contains a matrix object, "datExpr". Use data(Sample_Expression) to load.
}
\details{
a gene expression matrix.
}
\references{
1.	Song, W.M. and B. Zhang, Multiscale Embedded Gene Co-expression Network Analysis. PLoS Comput Biol, 2015. 11(11): p. e1004574.
}