\name{FrequenctSets.v6.2}
\docType{data}
\alias{FrequenctSets.v6.2.RData}
\title{
  Curated MSigDB signatures for testing funtion/pathway enrichments of MEGENA modules
}
\description{
  MSigDB v6.2 sets (BIOCARTA, GO, KEGG, REACTOME, HALLMARK and oncogenic signatures)
}
\usage{
  data(FrequenctSets.v6.2)
}
\format{
  Contains a list object, "msigdb.sets". Use data(Sample_Expression) to load.
}
\details{
  a list of curated gene signatures from MSigDB
}
