\name{MEGENA-package}
\alias{MEGENA-package}
\alias{MEGENA}
\docType{package}
\title{
co-expression network analysis
}
\description{
construction of gene-gene interaction network and dissection into multi-scale functional modules, and network key drivers. 
}
\details{
\tabular{ll}{
Package: \tab MEGENA\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2015-12-18\cr
License: \tab GPL (>= 2)\cr
}
Multiscale Embedded Gene Co-expression Network Analysis (MEGENA)
}
\author{
Won-Min Song

Maintainer: Won-Min Song <won-min.song@mssm.edu>
}
\references{
 Song W-M, Zhang B (2015) Multiscale Embedded Gene Co-expression Network Analysis. PLoS Comput Biol 11(11): e1004574. doi: 10.1371/journal.pcbi.1004574 
}
\keyword{ package }
\keyword{ network }
\keyword{ co-expression }
\keyword{ regulatory network }